Desktop/9t gen/requirements.txt
flask
twilio
openai
python-dotenv
pymongo